fix(frontend): correction CSS BacktestConfigPanel (scoped) feat(import): suppression option 'Planning Hebdo', remplacement automatique par défaut refactor(backend): import_calendar_files supprime automatiquement l'ancien calendrier du même nom

+++ src-tauri/src/schema.rs
@@ -33,9 +33,9 @@ diesel::table! {
         event_time -> Timestamp,
         impact -> Text,
         description -> Text,
-        actual -> Nullable<Float>,
-        forecast -> Nullable<Float>,
-        previous -> Nullable<Float>,
+        actual -> Nullable<Double>,
+        forecast -> Nullable<Double>,
+        previous -> Nullable<Double>,
         created_at -> Timestamp,
         calendar_import_id -> Integer,
     }
